Bad Practice - I made multiple different changes to the model in this commit.
Added significan neuron enhancements, added features to the neuron that follow the neuron model implemented in TN.

Added skeleton for new mapping (bitfield), but still using stock linear mapping.

Changed custom types from #MACROS to typedefs.

migrated all random rollback code to the model_main area, while moving state management into the neuron file

Changed the way leak and reset functions work, added a new calculation based leak function. I will need to replace the function pointer version with this to hopefully increase performance.

Changed the way that the random value (PRN, PRNTN) numbers for neurons are calculated - I just use the value as the maximum value for the neuron. Currently there is a catch if values are entered that are higher than the TN hardware can support, but if this catch is removed the model can support arbitrarially large random values

Cleaned up and renamed values inside neuron - hopefully fileds will be very clear as to thier type and equiv. in the original neuron model from the paper.

Found a mapping error, hopefully the changes to the neuron send code will fix this.

Model_main, when generating newurons, now implements a "simple" neuron, instead of a complex random neuron. "Simple" & "linear" neurons are used extensively in the TN architecture. This is a function call, so if a different neuron type is desired, it is easy enough to change it.

TimeTicks - I have changed the time of big ticks and little ticks as follows:
littleTick = 1 + jitter.
bigTick = 10  + jitter.

I did this to allow for parallel conservative simulation to be feasable. I will need to change bigTick to be NUM_NEURONS_PER_CORE + jitter for more accurate simulation.